+++ jax/experimental/jax2tf/__init__.py
@@ -13,5 +13,5 @@
 # limitations under the License.
 
 # flake8: noqa: F401
-from .jax2tf import convert, shape_as_value, split_to_logical_devices, PolyShape
+from .jax2tf import convert, dtype_of_val, shape_as_value, split_to_logical_devices, PolyShape
 from .call_tf import call_tf